grab cvars from settings

+++ code/Player/Player.Settings.cs
@@ -18,8 +18,8 @@ public class PlayerSettingsData
 	public bool ClassicExplosions { get; set; }
 	public bool ViewRoll { get; set; }
 	public string PlayerModel { get; set; }
-	public float PlayerModelColour1 { get; set; }
-	public float PlayerModelColour2 { get; set; }
+	public int PlayerModelColour1 { get; set; }
+	public int PlayerModelColour2 { get; set; }
 	public string SprayImage { get; set; }
 	public string SprayColour { get; set; }
 }
